'models_are_back'

+++ views/post-view/post_view.php
@@ -2,56 +2,40 @@
 $title = 'test';
 require_once __DIR__.'/../../templates/header.php';
 if (!isset($_GET['post_id'])) header('Location:'.$home);
+
+// $post_id is available in PostController
+require_once __DIR__.'/../../controllers/PostController.php';
 require_once __DIR__.'/../../controllers/CommentController.php';
 require_once __DIR__.'/../../controllers/LikeController.php';
-require_once __DIR__.'/../../controllers/PostController.php';
-
-// classes
-$posts = new PostController($pdo);
-$comment_controller = new Comment($pdo);
-$like = new Like($pdo);
-
-// need to be fixed because the user can enter invalid post id
-$post_id = $_GET['post_id'];
-$user_id = $_SESSION['user_id'] ?? null;
-$comments = $comment_controller->all_comments($post_id);
 
 
+$post_controller = new PostController;
+$comment_controller = new CommentController;
+$like_controller = new LikeController;
+$post = $post_controller->get_post(); 
+$comments = $comment_controller->all_comments();
 
+if ($_SERVER['REQUEST_METHOD'] === "POST"){
+    $method = htmlspecialchars($_POST['_method']);
+    $comment = htmlspecialchars($_POST['comment'] ?? null);
+    $comment_id = htmlspecialchars($_POST['comment_id'] ?? null);
 
-if ($_SERVER['REQUEST_METHOD'] === 'POST'){
-  $method = htmlspecialchars($_POST['_method']);
-  $new_comment = htmlspecialchars($_POST['comment'] ?? null) ;
-  $comment_id =  htmlspecialchars($_POST['comment_id'] ?? null);
-  //add comment to the post-------------------------------
-  if ($method ==='create_comment'){
-  $result = $comment_controller->create_comment($new_comment,$post_id);
+    if ($method === 'create_comment' && $comment){
 
-  }
-  //-------------------------------------------------------
-
-  // edit comment in the post------------------------------
-  if ($method === 'edit_comment'){
-    $comment_controller->edit_comment($new_comment,$comment_id);
-  }
-  //-------------------------------------------------------
-    // edit comment in the post----------------------------
-    if ($method === 'delete_comment'){
+      $comment_controller->create_comment($comment);
+    }
+    if ($method === 'edit_comment' && $comment){
+      $comment_controller->edit_comment($comment,$comment_id);
+    }
+    if ($method === 'delete_comment' && $comment_id){
       $comment_controller->delete_comment($comment_id);
     }
-  //-------------------------------------------------------
-  // add like----------------------------------------------
-    if ($method === 'add_like'){
-      $like->add_like_to_comment($comment_id,$post_id);
+    if ($method === 'add_like_to_comment' && $comment_id){
+      $like_controller->add_like_to_comment($comment_id);
     }
-  //-------------------------------------------------------
-
+    header("Location: $_SERVER[REQUEST_URI]");
 }
-// get data from controllers-------------------------------
-$post = $posts->get_post($post_id);
-$comments = $comment_controller->all_comments($post_id);
 
-//---------------------------------------------------------
 ?>
 <div class="d-flex m-3 justify-content-center">
     <div class="card " style="width:100%;text-align:center;">
@@ -70,6 +54,7 @@ $comments = $comment_controller->all_comments($post_id);
         <?php if ($is_logged_in) :?>
         <div data-mdb-input-init class="form-outline mb-4">
           <input type="hidden" name="_method" value="create_comment">
+          <input type="hidden" name="post_id" value="<?=$post_id ?>">
           <input type="text" name='comment' id="addANote" class="form-control" placeholder="Type comment..." />
           <button class="btn btn-success mt-1" type="submit">send</button>
         </div>
@@ -89,7 +74,7 @@ $comments = $comment_controller->all_comments($post_id);
               <div class="d-flex flex-row align-items-center">
                 <?php  if($is_logged_in): ?>
                 <form method="POST" class="m-2">
-                  <input type="hidden" name="_method" value="add_like">
+                <input type="hidden" name="_method" value="add_like_to_comment">
                   <input type="hidden" name="comment_id" value="<?= $comment['id']?>">
                   <button type="submit" class="small text-muted mb-0 btn btn-success btn-sm">Upvote?</button>
                 </form>
